Move Daemon to kovri-main.
