+++ src/app/DaemonWin32.cpp
