+++ src/app/Win32Service.cpp
