+++ src/app/Win32Service.h
