+++ src/client/CMakeLists.txt
@@ -9,7 +9,6 @@ set(CLIENT_SRC
   "AddressBook.cpp"
   "Destination.cpp"
   "ClientContext.cpp"
-  "Daemon.cpp"
   "I2PService.cpp"
   "util/Config.cpp"
   "util/Filesystem.cpp")
@@ -18,19 +17,6 @@ include_directories(
   "."
   "../"
   "../core/")
-
-if(CMAKE_SYSTEM_NAME STREQUAL "Linux")
-  list(APPEND CLIENT_SRC "DaemonLinux.cpp")
-elseif(CMAKE_SYSTEM_NAME STREQUAL "FreeBSD")
-  list(APPEND CLIENT_SRC "DaemonLinux.cpp")
-elseif(CMAKE_SYSTEM_NAME STREQUAL "Darwin")
-  list(APPEND CLIENT_SRC "DaemonLinux.cpp")
-elseif(CMAKE_SYSTEM_NAME STREQUAL "OpenBSD")
-  list(APPEND CLIENT_SRC "DaemonLinux.cpp")
-elseif(CMAKE_SYSTEM_NAME STREQUAL "Windows")
-  list(APPEND CLIENT_SRC "DaemonWin32.cpp")
-  list(APPEND CLIENT_SRC "Win32Service.cpp")
-endif()
 
 if(WITH_LIBRARY)
   add_library(${CLIENT_NAME} ${CLIENT_SRC})
